+++ src/web/decisionBar.tsx
@@ -1,7 +1,8 @@
-import React, { useRef, useEffect } from "react";
+import React, { useRef, useEffect, useState } from "react";
 import Button from "@mui/material/Button";
 import Grid from "@mui/material/Grid";
 import { reactionTimePassed } from "../utils/sharedFunctions";
+import { ButtonBaseActions } from "@mui/material";
 
 function selectOption(option: OptionT) {
   if (reactionTimePassed()) {
@@ -13,25 +14,30 @@ interface Props {
   options: OptionT[];
 }
 
-export const DecisionBar = (props: Props) => {
-  const buttonRef = useRef<HTMLButtonElement>(null);
+export const DecisionBar = ({ options }: Props) => {
+  const [focusedIndex, setFocusedIndex] = useState(-1);
+  const actionRefs = useRef<Array<ButtonBaseActions | null>>([]);
 
-  // Focus the first option
+  // Focus the first option, or the already focused one.
   useEffect(() => {
-    if (buttonRef.current) {
-      buttonRef.current.focus();
-    }
+    const indexToFocus = focusedIndex > -1 ? focusedIndex : 0;
+    actionRefs.current[indexToFocus]?.focusVisible();
   });
+
+  useEffect(() => {
+    actionRefs.current = actionRefs.current.slice(0, options.length);
+  }, [options]);
 
   return (
     <Grid container spacing={1}>
-      {props.options.map((option, index) => (
+      {options.map((option, index) => (
         <Grid item key={index}>
           <Button
             variant="outlined"
             color="primary"
             onClick={() => selectOption(option)}
-            ref={index === 0 ? buttonRef : null}
+            action={(actionObj) => (actionRefs.current[index] = actionObj)}
+            onFocus={() => setFocusedIndex(index)}
             sx={{ marginBottom: 1 }}
           >
             {option.label}